Add turbo-links form to project so it works with turboloinks-ios
I chose this apporach so that I could continue using server-side rendered
DOM to display form error messages and not have to subject myself to
the world of JavaScript.

Since I'm starting to get a lot of form helpers piling up, I created my
own method, `application_form`, that hides the complexity if `simple_form`,
`turbolinks_form`, and the rails `form_with` helper so that I can more
easily swap out the form implementation as the dust settles in the future.