Remove containers
This should simplify batch management for items since
only one model/tabel has to be manipulated. The thinking
is that a box is an item, and that item can contain other
items.

If I'm lucky, this commit will hopefully be "famous last words"
to a future contributor who will probably regret this fateful decision.

If I'm not lucky, nobody will see this because customers failed to
show up and make this a big enough deal to worry about.

We'll see! I just hope you're well versed in philosophy.