Raise an exception when appending children to an item when container=false
This prevents items that are not containers from accumulating child items.

I went down the route of creating a delegate for the association, as opposed
to using a `before_add` callback, because the closure_tree gem didn't make
it straight forward to add to the `children` association.

If I had more time I'd patch the closure_tree gem with these hooks, perhaps
`has_closure_tree before_add_child: :foo`, to the class configurator.

A validation is also included in item that prevents a user from changing
it from being a container with children, to not. The error will prompt them
to first remove all of the children before they can make it an item.