Refactor resources controllers to lazy load the resource and resources values
I ran into issues extending controllers, so I had to change the resource
and resources method do be a method that lazily calls find or scope. Before
this was loaded from an instance variable set by the callback. I ran into
issues because I had to deal with the sequence of when before_action filters
fire, which is hard to control.