Replace :id with :uuid param for Item resources
This makes it possible to have URLs that can be casually shared via
Messages with OpenGraph and display the contents of an Item in a preview
card.

To get this working in a sane way, I had to overhaul the AccountLayout
controller to work with the UuidField concern.

This commit also changes the URLs that are printed for labels. Instead of
just printing `/labels/:uuid`, the url `/labels/:uuid/scan` is utilized
to limit the impact of a future breaking change of switching the key for
the Label resource from :id to :uuid.