Replace root items creation with a router to the account
Creating items at this level doesn't make sense because its possible for
a person to pick a container from Account A and designate the item be
created in Account B.

Instead I just send the person to whatever account they want to create an
item. Since this isn't a common path (I'll probably delete it), I'm taking
the easy way out.