Add rooms to account creation
I'm not super happy about this yet, but it serves as a decent placeholder
to differentiate creating objects at the root account level over creating
them at the sub-item level.

We'll see how this shakes out; there's a chance that differentiating them
isn't a good idea at the controller level. It could just be a view thing.