Display unauthorized page when somebody accesses resources logged out.
This is needed to still load the open graph meta data that will be important
for when people share links to other people who might not be logged in.

This should only be needed for the Item model, so I kept the deny_access
logic mostly the same. If I find I'm doing this a lot I'd want to rethink
the approach.