Migrate items.shelf_life to just items.expires_at
The shelf life date range was more of a pain that it was worth, especially
when using with Chronic. It was setting ranges with Infinity values for
nils, which was a pain. I get sick of it when this code caused a bug when
copying items, which would have required even more handling.

Blurg.

Finally, and most importantly, people don't think of an expiration date
with a "whenver it started" date. What's that even called? To keep things
simpler I just removed it.