Replace email invitation flow with invitation link
This should be slightly more secure since it replaces an "email a link
that anybody can use to join the account" with a "anybody can request to
be on the account, but they need the account owners' approval" workflow.

This commit also adds Stimulus to the project, which is the most JS I care
to accumlate.