Remove "Labels" from account sidebar
I'm not a huge fan of this "feature" and I can't make sense of it, so I'm
removing it.

I'm also realzing that Templates for creating items is over-thinking it.
The sell should be that you can very quickly bang out a home inventory
by just slamming a bunch of items into a list of lists. Detail could be
added later.