Change "New" screens from selecting a template to just entering a name.
I didn't like all of the clicking and how long it took to select a template
before adding a thing. It just got in the way of the mental flow of adding
a bunch of things. Do you really want to be selecting between an "Item"
or a "Container" and a bunch of other stuff? Probably not. Its still there
if people need it, in the form of "Select from template..." that will go
to those screens.

We'll see what happens to those templates. Might remove them if people
are not using them.