Move `list-card--item` CSS into a view component
This feels a bit over-engineered at the moment, but does offer only a
slight advantage in terms of lines of code and organization. The thing I
learned doing this excercise is that it's not worth the effort of making
something that's a view component unless its more complex than these
cards.

I also learned about some of the limitations of view component composition,
mainly that inheritence doesn't really work with how the view component
APIs work.

I also learned about the `slots` API, which I think is a weird name. The
word was brought in from Vue/React world, but doesn't seem necessary. I'd
recommend to the crew to keep using the `with` blocks. There was also
some weird "magic" with the names of the method calls into slot collections.

From here, I might consider making the SearchField a component and the
Sidebar/Content layout a component.