Fix label scanning bugs
Most of these were caused by bugs that generated old URLs or used ids
in URLs that I'm trying to get rid of, like the long GUID and pk_id.

For now I just put a handler in the Labels controller that figures out
how to query the label from the format of the :id or :uuid param to look-up
the label. What I'd rather do is move this into a `/scans/v1/*` controller
so I can encode that into the QR code and better manage changes to URL
formats that end up being printed.