Add experimental a way to launch items, people, and profile from a URL
Since people can belong to multiple accounts, I need a way to peg to one
account on mobile devices so the tab navigations works correctly and opens
the correct account when somebody selects "Items", "People", or "Profile".

I'm trying to do this via cookies set by the server so that complexity
does not have to live within the client-side application.