Split the SvgFile and Icon concerns to handle more generic SVG assets
I needed to render SVG illustrations for the content site and not have
them appear under icons, so I started splitting those concerns. I'm not
too pleased with how its turned out; there's still a lot of concerting
from an object to a string back to an object that doesn't feel right and
there are some things I think that could be simplified for the component.