Remove redirect logic from item creation
It was just kinda of weird; most of the time I found myself wanting to
go into the item to edit the thing, and it was hard always trying to find
it from the list after I added it.

I could add a "View item" link to the flash notice, but not now.