Email people who create new home accounts
This is setup to work for other types of accounts too from the other
templates. For now I just got homes working because that's the only
signups so far.

I also updates the signature, layouts, and a few other emails so that
email works better overall.

Oh, and I threw in a few email previews so its easy to read the email
content for development.