Move view code that generates PDFs from Controller into .ruby files.
Wait, this is kind of crazy. In Rails, if you have code that renders PDFs like
prawn, you don't have to stick all of that into a controller just because it's ruby.

What you can do is add a `show.pdf.ruby` file as a view. That view will
run whatever code you put in there and render the output of the file. Whaaa?

Make sure whatever view code you come up with in your view is some strict DSL
that actually lays out the view. If you're putting business logic in there you'll be asking
for pain!

Prawn is a good DSL that limits the business logic in what ultimately just renders a PDF.
If you think about it, template languages like HAML are DSLs, so why not run a ruby DSL?

I'll dig into the guts of the Rails renderer and create a `show.prawn.pdf` handler that makes
sure it runs only the Prawn DSL. This could be released as a Gem and we'd save the world
from the oppression of DSL rendering in a controller.