Move PDF rendering from Client to Server
Why? Grab a chair.

If I try to show a preview of the PDF in the html page, the browsers stop responding because
it leaks memory into the page that's managed by Turbolinks (doesn't real). There were also some
Z-ordering issues where the embed controls would float controls above the drop-down menu. It
just wouldn't work.

I initially thought that I should use PDF.js to render that stuff client side. It could be managed by
Turbolinks and cleaned up properly, right? That lead into an insane rabbit hole down the belly of
webpack, yarn, npm ... or whatever they call it now. It would have been janky.

I felt defeated.

But then I thought, "what if I just render the PDF into a PNG and show it client-side via an `img`
tag. It worked! And all browsers can handle it. And it gets cleaned up. And I can optimize it so much
more on the server and got a lot of mileage out of it.

I did it and it worked. So here we are. Proof that you should probably do as much on the server as
possible.