Deprecate label.uuid by replacing with label.code
Confirmed! The labels schema and Label model are worthless, so this starts
the beginning of deprecation. This commit makes sure that hand written codes
start using the UUID derivation of the Item (instead of the Label). It
accomplishes this by delegating the base UUID to the Item and changes
the search UUID to look at the item instead of the label.

Next step will be removing labels, which will require a migration of the
label UUIDs into the item UUIDs, then deleting labels schema and replacing
the Label ActiveRecord object with an ActiveModel that delegates to Item.