Got a stripe-hosted checkout page working
Its very basic, but that is the point. I can now collect dollars for a single plan,
but it does not yet process a callback to update the standing of the account. That's
fine because accounts are fully unlocked.

This will obviously be enhanced over time to have stronger controls that protect
non-upgraded account resources.