Closes issue at https://github.com/thingybase/community/issues/18
The user was very specific about how search should work, which made sense

I'm not sure if I'm doing this right ... I'm pretty sure I need to add a GIN index
to the items.name column to make it run quickly. Since the datasets are small, I'm
not going to worry too much about that, but I should wire up APM to this so I can get
slow query alerts.